+++ src/o_init.c
@@ -30,7 +30,7 @@ STATIC_OVL void
 shuffle_tiles()
 {
     int i;
-    short tmp_tilemap[NUM_OBJECTS];
+    int tmp_tilemap[NUM_OBJECTS];
 
     for (i = 0; i < NUM_OBJECTS; i++)
         tmp_tilemap[i] = glyph2tile[objects[i].oc_descr_idx + GLYPH_OBJ_OFF];
