+++ win/win32/sdl/sdlproc.c
@@ -2319,7 +2319,7 @@ sdl_initMapTiles(void)
     DWORD errcode;
     int tl_num;
     SIZE map_size;
-    int total_tiles_used = process_tiledata(2, (const char*)0, (short*)0, (uchar*)0);
+    int total_tiles_used = process_tiledata(2, (const char*)0, (int*)0, (uchar*)0);
 
     /* no file - no tile */
     if (!(iflags.wc_tile_file && *iflags.wc_tile_file))
